UIFC-344-address-code-duplications useUpdatedFilters
